Make another constructor that only takes the hashed elements
This is a change to allow elements to be added to the tree, without the
need for the original elements to be added to the tree. Instead, one can
throw away the current tree and using the hashes stored in the leaves
plus the new element to be added and build a new merkle tree.