remove ignore macro usages and fix tree to make tests pass
Also some tests were fixed. First by using the appropiate names (some
tests that added elements named the variables with names implying
something was removed). Then some expected node counts were wrong. This
was because the cases we want to check are:

- having 2**k elements and adding stuff
- having (2**k) + 1 elements and adding stuff
- having 2**k elements and removing stuff
- having (2**k) + 1 elements and removing stuff

also, when the input elements had a size of (2**k) + 1, the actual
amount of leaves becomes 2**(k+1) and the node count is 2**(k+1) * 2
(the *2 is missing)